+++ src/math_utils.h
@@ -7,14 +7,14 @@
 namespace uwmf
 {
 
-float mean(const monochrome_image& image);
-float variance(const monochrome_image& image, const float m);
-float variance(const monochrome_image& image);
-float covariance(const monochrome_image& image1,
+double mean(const monochrome_image& image);
+double variance(const monochrome_image& image, const double m);
+double variance(const monochrome_image& image);
+double covariance(const monochrome_image& image1,
         const monochrome_image& image2);
-float covariance(const monochrome_image& image1, const float v1,
-        const monochrome_image& image2, const float v2);
-float se(const monochrome_image& image1, const monochrome_image& image2);
-float mse(const monochrome_image& image1, const monochrome_image& image2);
+double covariance(const monochrome_image& image1, const double v1,
+        const monochrome_image& image2, const double v2);
+double se(const monochrome_image& image1, const monochrome_image& image2);
+double mse(const monochrome_image& image1, const monochrome_image& image2);
 
 } // uwmf
